Fixing bug with filters for asset receiving status that was affecting pagination.
The bug was that the back end index.py was forming and sending the wrong query parameters in the JSON response, in the form of 'asset.is_current__includes=[<0 (false) or 1 (true) for each status type>]' instead of the correct form, 'past=true|false?present=true|false?future=true|false'.  TODO: Refactor logic from index.py that takes query parameters to determine (1) filters to be used in response data for filters (query params) appended to pagination links and separately, (2) formation of part of the syntax to be used to help form the database queries that should filter the asset listing in the SQL WHERE clause.